requirements.txt
google-generativeai==0.5.2
python-dotenv
twstock
pandas
matplotlib
cloudinary
flask
line-bot-sdk==2.4.1
requests
newspaper3k
